pyproject.toml
[project]
name = "gaussianspec"
version = "0.1.0"
description = "Gaussian elimination specification"
readme = "README.md"
authors = [{ name = "Alok Singh", email = "alokbeniwal@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "morphcloud>=0.1.67",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "pantograph @ git+https://github.com/stanford-centaur/PyPantograph.git@main",
    "gymnasium>=0.29.1",
    "torch>=2.3.0",
    "transformers>=4.41.0",
    "python-dotenv>=1.0.1",
    "huggingface_hub>=0.23.0",
    "accelerate>=0.30.0",
    "stable-baselines3>=2.3.0",
    "pytest>=8.2.1",
    "google-generativeai>=0.8.5",
    "openai>=1.30.0",
    "tqdm>=4.66.4",
    "httpx>=0.28.0",
    "jaxtyping>=0.3.2",
    "pypdf2>=3.0.1",
    "pdfminer-six>=20250416",
    "reportlab>=4.4.0",
    "litellm>=1.34.10",
]

[project.scripts]
gaussianspec = "gaussianspec:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ipykernel>=6.29.5", "ruff>=0.11.9"]
[tool.ruff.format]
docstring-code-format = true

[tool.hatch.metadata]
allow-direct-references = true
